Extract feature data and icons * Move features data and management to data-stores * Move icons to onboarding package
