+++ client/landing/gutenboarding/hooks/use-recommended-plan.ts
@@ -7,9 +7,8 @@ import { Plans } from '@automattic/data-stores';
 /**
  * Internal dependencies
  */
-import { STORE_KEY as ONBOARD_STORE } from '../stores/onboard';
 import { PLANS_STORE } from '../stores/plans';
-import { FEATURE_LIST } from '../onboarding-block/features/data';
+import { WPCOM_FEATURES_STORE } from '../stores/wpcom-features';
 
 const order = [
 	Plans.PLAN_PERSONAL,
@@ -19,16 +18,19 @@ const order = [
 ];
 
 export function useRecommendedPlanSlug() {
-	const selectedFeatures = useSelect( ( select ) => select( ONBOARD_STORE ).getSelectedFeatures() );
+	const allFeatures = useSelect( ( select ) => select( WPCOM_FEATURES_STORE ).getAllFeatures() );
+	const selectedFeatures = useSelect( ( select ) =>
+		select( WPCOM_FEATURES_STORE ).getSelectedFeatures()
+	);
 
 	if ( ! selectedFeatures.length ) return;
 
 	return selectedFeatures.reduce( ( currentMinSupportedPlan, featureId ) => {
-		const featureMinSupportedPlan = FEATURE_LIST[ featureId ].minSupportedPlan;
+		const featureMinSupportedPlan = allFeatures[ featureId ].minSupportedPlan;
 		return order.indexOf( featureMinSupportedPlan ) > order.indexOf( currentMinSupportedPlan )
 			? featureMinSupportedPlan
 			: currentMinSupportedPlan;
-	}, FEATURE_LIST[ selectedFeatures[ 0 ] ].minSupportedPlan );
+	}, allFeatures[ selectedFeatures[ 0 ] ].minSupportedPlan );
 }
 
 export default function useRecommendedPlan() {
