+++ client/landing/gutenboarding/onboarding-block/features/index.tsx
@@ -5,8 +5,9 @@ import * as React from 'react';
 import { useDispatch, useSelect } from '@wordpress/data';
 import { useI18n } from '@automattic/react-i18n';
 import { Button } from '@wordpress/components';
-import { Icon } from '@wordpress/icons';
+import type { WPCOMFeatures } from '@automattic/data-stores'
 import {
+	FeatureIcon,
 	Title,
 	SubTitle,
 	ActionButtons,
@@ -19,8 +20,7 @@ import classnames from 'classnames';
 /**
  * Internal dependencies
  */
-import { STORE_KEY as ONBOARD_STORE } from '../../stores/onboard';
-import { FEATURE_LIST, FeatureId } from './data';
+import { WPCOM_FEATURES_STORE } from '../../stores/wpcom-features';
 import useStepNavigation from '../../hooks/use-step-navigation';
 
 /**
@@ -28,12 +28,18 @@ import useStepNavigation from '../../hooks/use-step-navigation';
  */
 import './style.scss';
 
+type FeatureId = WPCOMFeatures.FeatureId;
+
 const FeaturesStep: React.FunctionComponent = () => {
 	const { __ } = useI18n();
 	const { goBack, goNext } = useStepNavigation();
 
-	const selectedFeatures = useSelect( ( select ) => select( ONBOARD_STORE ).getSelectedFeatures() );
-	const { addFeature, removeFeature } = useDispatch( ONBOARD_STORE );
+	const allFeatures = useSelect( ( select ) => select( WPCOM_FEATURES_STORE ).getAllFeatures() );
+
+	const selectedFeatures = useSelect( ( select ) =>
+		select( WPCOM_FEATURES_STORE ).getSelectedFeatures()
+	);
+	const { addFeature, removeFeature } = useDispatch( WPCOM_FEATURES_STORE );
 
 	const hasSelectedFeatures = selectedFeatures.length > 0;
 
@@ -67,7 +73,7 @@ const FeaturesStep: React.FunctionComponent = () => {
 			</div>
 			<div className="features__body">
 				<div className="features__items">
-					{ Object.entries( FEATURE_LIST ).map( ( [ id, feature ] ) => (
+					{ Object.entries( allFeatures ).map( ( [ id, feature ] ) => (
 						<Button
 							className={ classnames( 'features__item', {
 								'is-selected': selectedFeatures.includes( feature.id ),
@@ -77,7 +83,7 @@ const FeaturesStep: React.FunctionComponent = () => {
 							isTertiary
 						>
 							<div className="features__item-image">
-								<Icon icon={ feature.icon } />
+								<FeatureIcon featureId={ feature.id } />
 							</div>
 							<div className="features__item-heading">
 								<div className="features__item-name">{ feature.name }</div>
