+++ client/landing/gutenboarding/stores/onboard/actions.ts
@@ -15,7 +15,6 @@ import { SITE_STORE } from '../site';
 import { PLANS_STORE } from '../plans';
 import type { State } from '.';
 import type { FontPair } from '../../constants';
-import type { FeatureId } from '../../onboarding-block/features/data';
 import { isEnabled } from 'config';
 
 type CreateSiteParams = Site.CreateSiteParams;
@@ -24,11 +23,6 @@ type Template = VerticalsTemplates.Template;
 type Language = {
 	value: number;
 };
-
-export const addFeature = ( featureId: FeatureId ) => ( {
-	type: 'ADD_FEATURE' as const,
-	featureId,
-} );
 
 export function* createSite(
 	username: string,
@@ -93,9 +87,8 @@ export function* createSite(
 	return success;
 }
 
-export const removeFeature = ( featureId: FeatureId ) => ( {
-	type: 'REMOVE_FEATURE' as const,
-	featureId,
+export const enableExperimental = () => ( {
+	type: 'SET_ENABLE_EXPERIMENTAL' as const,
 } );
 
 export const resetFonts = () => ( {
@@ -201,8 +194,7 @@ export const startOnboarding = () => ( {
 } );
 
 export type OnboardAction = ReturnType<
-	| typeof addFeature
-	| typeof removeFeature
+	| typeof enableExperimental
 	| typeof resetFonts
 	| typeof resetOnboardStore
 	| typeof resetSiteVertical
