+++ client/landing/gutenboarding/stores/onboard/index.ts
@@ -34,7 +34,6 @@ registerStore< State >( STORE_KEY, {
 		'plan',
 		'randomizedDesigns',
 		'selectedDesign',
-		'selectedFeatures',
 		'selectedFonts',
 		'selectedSite',
 		'siteTitle',
