+++ client/landing/gutenboarding/stores/onboard/reducer.ts
@@ -11,7 +11,6 @@ import type { DomainSuggestions, Plans } from '@automattic/data-stores';
 import type { SiteVertical, Design } from './types';
 import type { OnboardAction } from './actions';
 import type { FontPair } from '../../constants';
-import type { FeatureId } from '../../onboarding-block/features/data';
 
 // Returns true if the url has a `?latest`, which is used to enable experimental features
 export function hasExperimentalQueryParam() {
@@ -153,25 +152,6 @@ const selectedDesign: Reducer< Design | undefined, OnboardAction > = ( state, ac
 	return state;
 };
 
-const selectedFeatures: Reducer< FeatureId[], OnboardAction > = (
-	state: FeatureId[] = [],
-	action
-) => {
-	if ( action.type === 'ADD_FEATURE' ) {
-		return [ ...state, action.featureId ];
-	}
-
-	if ( action.type === 'REMOVE_FEATURE' ) {
-		return state.filter( ( id ) => id !== action.featureId );
-	}
-
-	if ( action.type === 'RESET_ONBOARD_STORE' ) {
-		return [];
-	}
-
-	return state;
-};
-
 const selectedSite: Reducer< number | undefined, OnboardAction > = (
 	state = undefined,
 	action
@@ -250,7 +230,6 @@ const reducer = combineReducers( {
 	siteVertical,
 	showSignupDialog,
 	plan,
-	selectedFeatures,
 	wasVerticalSkipped,
 	isExperimental,
 	randomizedDesigns,
