+++ client/landing/gutenboarding/stores/wpcom-features/index.ts
@@ -1,0 +1,6 @@
+/**
+ * External dependencies
+ */
+import { WPCOMFeatures } from '@automattic/data-stores';
+
+export const WPCOM_FEATURES_STORE = WPCOMFeatures.register();
