+++ packages/data-stores/src/index.ts
@@ -8,6 +8,7 @@ import persistenceConfigFactory from './persistence-config-factory';
 import * as Plans from './plans';
 import * as Site from './site';
 import * as Verticals from './verticals';
+import * as WPCOMFeatures from './wpcom-features';
 import * as VerticalsTemplates from './verticals-templates';
 
 export {
@@ -18,6 +19,7 @@ export {
 	Verticals,
 	VerticalsTemplates,
 	Plans,
+	WPCOMFeatures,
 	persistenceConfigFactory,
 };
