+++ packages/data-stores/src/wpcom-features/constants.tsx
@@ -1,0 +1,1 @@
+export const STORE_KEY = 'automattic/wpcom-features';
