+++ packages/onboarding/package.json
@@ -31,6 +31,7 @@
 	},
 	"dependencies": {
 		"@automattic/react-i18n": "^1.0.0-alpha.0",
+		"@automattic/data-stores": "^1.0.0-alpha.1",
 		"@wordpress/components": "^10.0.5",
 		"classnames": "^2.2.6"
 	},
