+++ packages/onboarding/src/index.ts
@@ -1,2 +1,3 @@
 export { Title, SubTitle } from './titles';
 export { default as ActionButtons, BackButton, NextButton, SkipButton } from './action-buttons';
+export { default as FeatureIcon } from './feature-icon';
